% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hallo}
\alias{hallo}
\alias{halloechen}
\title{Hallo function}
\usage{
hallo(txt = "world")

halloechen(txt = "world")
}
\arguments{
\item{txt}{\code{character} vector}
}
\description{
Hallo function
}
\details{
More informations.
}
\seealso{
\code{\link[base]{cat}}
}
\author{
Michal
}
