% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello function}
\usage{
hello(txt = "world")
}
\arguments{
\item{txt}{\code{character} vector}
}
\description{
Hello function
}
\details{
More informations.
}
\seealso{
\code{\link[base]{cat}}
}
\author{
Michal
}
