% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_help.R
\name{pyhelp}
\alias{pyhelp}
\title{Simple Python topic help}
\usage{
pyhelp(topic)
}
\arguments{
\item{topic}{character-valued Python help topic}
}
\value{
Python help text
}
\description{
Equivalent to \code{help(topic)} issued from Python.
}
\examples{
pyhelp("numpy.linalg.svd")
}

