% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{R}
\alias{R}
\title{Convert Python objects to R}
\usage{
R(x)
}
\arguments{
\item{x}{a Python object}
}
\value{
a converted R value
}
\description{
The python package defers conversion of Python objects into R
objects in order to allow composition of Python objects. Use
the \code{R} function to convert a Python object into an R
value.
}
\note{
The \code{R} function copies data from Python to R.
}
\examples{
\dontrun{
np <- import("numpy")
x <- matrix(1:9, 3)
p <- np$matrix(x)
p$sum()     # Python scalar value
R(p$sum())  # R scalar value

l <- np$linalg$eig(p)  # Python list
# compare with R's eigen function
eigen(x)$values - R(l)[[1]]
}

}

