% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{r_py_to_r}
\alias{r_py_to_r}
\title{Convert miscellaneous Python objects to R objects called by the
py_to_r C function as a last conversion result}
\usage{
r_py_to_r(x)
}
\arguments{
\item{x}{a python object}
}
\value{
either a converted R object or NULL if no
         suitable conversion could be found
}
\description{
Convert miscellaneous Python objects to R objects called by the
py_to_r C function as a last conversion result
}
\keyword{internal}

