% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{r_r_to_py}
\alias{r_r_to_py}
\title{Convert miscellaneous R objects to Python objects, called by r_to_py
C function as a last conversion result}
\usage{
r_r_to_py(x)
}
\arguments{
\item{x}{an R object}
}
\value{
either a converted Python object or error if no
         suitable conversion could be found
}
\description{
Convert miscellaneous R objects to Python objects, called by r_to_py
C function as a last conversion result
}
\keyword{internal}

